ref(test): Split setup into global (DB seed) and auth (browser login) projects
