+++ tests/global.setup.ts
@@ -1,15 +1,8 @@
-import path from 'node:path';
-import { expect, test as setup } from '@playwright/test';
+import { test as setup } from '@playwright/test';
 import { createClient } from '@supabase/supabase-js';
+import { TEST_USER } from './test-user';
 
-const TEST_USER = {
-  email: 'e2e-test@hibana.com',
-  password: 'e2e-test-password',
-};
-
-export const AUTH_FILE = path.join(__dirname, '../playwright/.auth/user.json');
-
-setup('create test user and authenticate', async ({ page }) => {
+setup('seed test user', async () => {
   const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
   const serviceKey = process.env.SUPABASE_SECRET_KEY;
 
@@ -46,17 +39,4 @@ setup('create test user and authenticate', async ({ page }) => {
   if (createError) {
     throw new Error(`Failed to create test user: ${createError.message}`);
   }
-
-  // Log in via the browser UI to get proper auth cookies
-  await page.goto('/login');
-
-  await page.getByLabel(/email/i).fill(TEST_USER.email);
-  await page.getByLabel(/password/i).fill(TEST_USER.password);
-  await page.getByRole('button', { name: /log in/i }).click();
-
-  // Wait for redirect to dashboard after successful login
-  await expect(page).toHaveURL('/dashboard', { timeout: 15_000 });
-
-  // Save authenticated browser state
-  await page.context().storageState({ path: AUTH_FILE });
 });
